refactor: 도메인 로직과 UI 로직 폴더 분리하기
